video from image beta

+++ src/main/kotlin/ru/tcloud/vmaker/core/command/impl/CreateFromDirCommand.kt
@@ -5,39 +5,46 @@ import ru.tcloud.vmaker.core.command.Command
 import ru.tcloud.vmaker.core.command.CommandType
 import ru.tcloud.vmaker.core.editor.Editor
 import ru.tcloud.vmaker.core.exception.VMakerException
-import ru.tcloud.vmaker.core.support.CircleCollection
+import ru.tcloud.vmaker.core.worker.AudioWorker
+import ru.tcloud.vmaker.core.worker.FileDescription
+import ru.tcloud.vmaker.core.worker.ImageWorker
+import ru.tcloud.vmaker.core.worker.VideoWorker
+import ru.tcloud.vmaker.core.worker.VideoWorker.Codec.LIBX264
+import ru.tcloud.vmaker.core.worker.VideoWorker.VideoExtension.MP4
 import java.io.File
+import java.util.*
+
+//cfd -w /home/evgeny/Documents/Chillax/video1 -d 30
 
 @Service
-class CreateFromDirCommand(private val editor: Editor) : Command {
+class CreateFromDirCommand(
+    private val videoWorker: VideoWorker,
+    private val audioWorker: AudioWorker,
+    private val imageWorker: ImageWorker
+    ) : Command {
 
     private enum class Arguments(val arg: String, val desc: String) {
         W(
             "w", """work dir, have to follow next rule workDir
             |                                                   - video
             |                                                   - music
+            |                                                   - picture
         """.trimMargin()
         ),
         D("d", "video have to be duration in min"),
-        C("c", "clean temp dir after work (true/false), default false"),
+        C("c", "clean temp dir after work (true/false), default $cleanAfterDefaultValue"),
+        P("t", "picture to video, use -p 10 where 1- is duration in sec, default ${pictureToVideoDefaultDuration}s"),
         HELP("help", """"${W.arg} - ${W.desc}
             |                ${C.arg} - ${C.desc}
             |                ${D.arg} - ${D.desc}
+            |                ${P.arg} - ${P.desc}
             |                """
             .trimMargin())
     }
     private val videoDirName = "video"
     private val musicDirName = "music"
+    private val pictureDirName = "picture"
     private val dirStructure = setOf(videoDirName, musicDirName)
-    private val mp4 = "mp4"
-    private val mp3 = "mp3"
-
-    private fun argToEnum(str: String): Map<String, Arguments> {
-        return Arguments.values().associateBy { it.desc }
-    }
-
-    private data class TmpFilesHolder(val fadeFile: File, val codecFile: File)
-    private data class FileDescription(var duration: Int, var tmpFiles: TmpFilesHolder? = null)
 
     override fun getType() = CommandType.CREATE_FROM_DIR
 
@@ -47,43 +54,57 @@ class CreateFromDirCommand(private val editor: Editor) : Command {
             return
         }
         val workDirPath = args[Arguments.W.arg]?: throw VMakerException("You need set -w flag, use -help for more info")
-        var duration = (args[Arguments.D.arg]?: throw VMakerException("You need set -d flag, use -help for more info")).toInt() * 60
-        val cleanTemp = args[Arguments.C.arg]?: true
+        val duration = (args[Arguments.D.arg]?: throw VMakerException("You need set -d flag, use -help for more info")).toInt() * 60
+        val cleanTemp = if(args.containsKey(Arguments.C.arg)) args[Arguments.C.arg].toBoolean() else cleanAfterDefaultValue
+        val pictureToVideoDuration = if(args.containsKey(Arguments.P.arg)) args[Arguments.P.arg]?.toInt() else pictureToVideoDefaultDuration
+
         checkWorkDir(workDirPath)
-        var workDir = File("${workDirPath.trim()}")
+        val workDir = File(workDirPath.trim())
         val videoDir = File("${workDirPath.trim()}/$videoDirName")
         val musicDir = File("${workDirPath.trim()}/$musicDirName")
+        val pictureDir = File("${workDirPath.trim()}/$pictureDirName")
         val tmpFiles = mutableSetOf<File>()
         try {
-            val audiFiles = fillAudioToFinallyResult(duration, musicDir)
-            val concatenationMp3 = if(audiFiles.size > 1) editor.concatenationMp3(musicDir, audiFiles, tmpFiles) else audiFiles[0]
-            duration = editor.getDuration(concatenationMp3)
-            val prepareVideoToConcat = prepareVideoToConcat(workDirPath)
-                .onEach { (t, _) -> tmpFiles.add(t) }
-            val concatVideo = if(prepareVideoToConcat.map { it.value.duration }.sum() >= duration.toLong()) {
-                prepareVideoToConcat.mapNotNull { it.value.tmpFiles?.codecFile }
-            } else {
-                fillVideoToFinallyResult(duration, prepareVideoToConcat.entries.map { it.value })
-            }
-            val concatenation = editor.concatenationMp4(videoDir, concatVideo).apply { tmpFiles.add(this) }
-            val silent = editor.doSilent(videoDir, concatenation).apply { tmpFiles.add(this) }
-            editor.addAudioOnVideo(concatenationMp3, silent, workDir)
+            val concatenationMp3 = audioWorker
+                .fillAudioToFinallyResult(duration, musicDir)
+                .let {
+                    audioWorker.concatenationMp3(musicDir, it, tmpFiles)
+                }
+            val fadeVideos = prepareVideoToConcat(videoDir, pictureDir, tmpFiles, pictureToVideoDuration?:pictureToVideoDefaultDuration)
+            val concatVideo = videoWorker.fillVideoToFinallyResult(concatenationMp3.duration, fadeVideos)
+            val concatenation = videoWorker.concatenationMp4(videoDir, concatVideo, tmpFiles).apply { tmpFiles.add(this) }
+            val silent = videoWorker.doSilent(videoDir, concatenation, tmpFiles).apply { tmpFiles.add(this) }
+            videoWorker.addAudioOnVideo(concatenationMp3.file, silent, workDir, tmpFiles)
         } finally {
-            tmpFiles.forEach { it.delete() }
+            if(cleanTemp) {
+                tmpFiles.forEach { it.delete() }
+            }
         }
     }
 
-    private fun prepareVideoToConcat(workDirPath: String): Map<File, FileDescription> {
-        val videos = File(workDirPath.trim()).listFiles()?.find { it.name == videoDirName } ?: throw VMakerException(Arguments.W.desc)
+    private fun prepareVideoToConcat(videoDir: File, imageDir: File, tmpFiles: MutableSet<File>, duration: Int): List<FileDescription> {
+        return fromVideo(videoDir, tmpFiles)
+            .plus(fromImages(imageDir, tmpFiles, duration))
+    }
+
+    private fun fromVideo(videoDir: File, tmpFiles: MutableSet<File>): List<FileDescription> {
+        val videos = videoDir.listFiles()?.filter { it.extension == MP4.ex }.orEmpty()
         var counter = 1
-        return videos.listFiles()
-            .filter { it.extension == mp4 }.associateWith {
-                val duration = editor.getDuration(it)
-                val fadeVideo = editor.makeFadeVideo(counter, it, duration)
-                val encodeVideo = editor.encodeVideo(counter, fadeVideo)
-                counter++
-                FileDescription(duration, TmpFilesHolder(fadeVideo, encodeVideo))
-            }
+        return videos.map {
+            videoWorker.makeFadeVideo("$counter", it, tmpFiles)
+                .apply { counter++ }
+        }
+    }
+
+    private fun fromImages(imageDir: File, tmpFiles: MutableSet<File>, duration: Int): List<FileDescription> {
+        val img = imageDir.listFiles()
+            ?.filter { ImageWorker.ImageExtension.values().map { v -> v.name.lowercase(Locale.getDefault()) }
+                .contains(it.extension) }.orEmpty()
+        var counter = 1
+        return img.map {
+            FileDescription(duration, imageWorker.imageToFadeVideo(it, "$counter", duration, tmpFiles))
+                .apply { counter++ }
+        }
     }
 
     private fun checkWorkDir(workDirPath: String) {
@@ -96,41 +117,9 @@ class CreateFromDirCommand(private val editor: Editor) : Command {
         }
     }
 
-    private fun fillVideoToFinallyResult(totalDuration: Int, src: List<FileDescription>): List<File> {
-        var durationNow = totalDuration
-        val circleCollection = CircleCollection(src.map { it  })
-        val result = mutableListOf<File>()
-        while (durationNow >= 0) {
-            val next = circleCollection.next()
-            durationNow -= next.duration
-            result.add(next.tmpFiles?.codecFile?:throw IllegalArgumentException("File can't be null"))
-        }
-        return result
+    companion object {
+        internal const val pictureToVideoDefaultDuration = 30
+        internal const val cleanAfterDefaultValue = false
     }
-
-    private fun fillAudioToFinallyResult(totalDuration: Int, audioDir: File): List<File> {
-        val mp3s = audioDir.listFiles()
-            .apply {
-                if(this == null || this.isEmpty()) {
-                    throw VMakerException("Can't find in ${audioDir.absolutePath}")
-                }
-            }
-            ?.filter { it.extension == mp3 }
-            ?.associate { it to editor.getDuration(it) }?.entries
-        if(mp3s == null || mp3s.isEmpty()) {
-            throw VMakerException("$audioDir is doesn't content MP3")
-        }
-        var durationNow = totalDuration
-        val circleCollection = CircleCollection(mp3s)
-        val result = mutableListOf<File>()
-        while (durationNow >= 0) {
-            val next = circleCollection.next()
-            durationNow -= next.value
-            result.add(next.key)
-        }
-        return result
-    }
-
-    private fun minToSec(min: Int): Int = min * 60
 
 }
